fix: resolve all clippy warnings for Rust best practices
- Fix bool_assert_comparison: use assert!(x) instead of assert_eq!(x, true)
- Fix redundant_closure: use Vec::new instead of || Vec::new()
- Fix len_zero: use !is_empty() instead of len() > 0
- Fix syntax errors in application_engine_tests.rs
- Update stack_item tests for idiomatic assertions

+++ crates/vm/tests/csharp_script_builder_tests.rs
@@ -385,7 +385,7 @@ fn expected_jump_sequence(opcode: OpCode, offset_i8: i32, offset_i32: i32) -> Ve
         result.extend_from_slice(&offset_i32.to_le_bytes());
     } else {
         result.push(code);
-        result.extend_from_slice(&(offset_i8 as i32).to_le_bytes());
+        result.extend_from_slice(&offset_i8.to_le_bytes());
         result.push(code);
         result.extend_from_slice(&offset_i32.to_le_bytes());
     }
